Fix building packages that use autoconf on PIE-by-default systems
This changes "-fno-PIE" to "-no-pie" in the options passed to GCC when compiling
C code. That makes packages that invoke GCC in their build systems - e.g.
anything that uses autoconf - not fail. Details:

"-fno-PIE" affects compilation only, while "-no-pie" affects compilation and
linking. GHC always invokes GCC's linking and compilation facilities separately,
so this isn't a problem. But if a package attempts to build an executable in one
shot with GCC, it will fail. Autoconf does this for feature testing, and if a
package included an executable written in C that would fail too.

Tested by: modifying ~/.stack/config.yaml to point to the modified one, then
building network in the global package. Works with the change, fails without.